+++ front_end/package.json
@@ -12,9 +12,11 @@
     "@material-ui/icons": "^4.11.2",
     "@types/node": "^14.14.25",
     "@types/react": "^17.0.1",
+    "formik": "^2.2.6",
     "next": "10.0.6",
     "react": "17.0.1",
     "react-dom": "17.0.1",
-    "typescript": "^4.1.3"
+    "typescript": "^4.1.3",
+    "yup": "^0.32.9"
   }
 }
